Add CSV option, remove "annotate" feature
This patch adds an option to append found keys & params into a csv file. This can be used to collect all error messages in a project.

Also, remove the "a" parameter which allowed to use a custom annotation for errror structs. This was too much and not really useful as we always want the json key to be the lower case version of the struct member.

Also, use the new "embed" stdlib feature to embed standard templates. Removes the "esc" package which was used to include binary data in the Go build.

Bumped Go version to 1.20.